user registration with hashed pw
users can now register with a hashed password being saved to the db
