+++ package.json
@@ -19,7 +19,6 @@
     "next": "^14.0.2",
     "next-auth": "^5.0.0-beta.18",
     "postcss": "8.4.31",
-    "prisma": "^5.13.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-swipeable": "^7.0.1",
@@ -39,7 +38,8 @@
     "eslint-config-next": "^14.0.0",
     "eslint-config-prettier": "9.0.0",
     "prettier": "^3.0.3",
-    "prettier-plugin-tailwindcss": "0.5.4"
+    "prettier-plugin-tailwindcss": "0.5.4",
+    "prisma": "^5.13.0"
   },
   "engines": {
     "node": ">=18.17.0"
